+++ tests/test_guinea_pig.py
@@ -1,56 +1,5 @@
 import pytest
-from os import environ
 
-from selenium import webdriver
-from selenium.common.exceptions import WebDriverException
-from selenium.webdriver.remote.remote_connection import RemoteConnection
-
-
-@pytest.fixture(scope='function')
-def driver(request):
-    # if the assignment below does not make sense to you please read up on object assignments.
-    # The point is to make a copy and not mess with the original test spec.
-    desired_caps = {}
-    
-    browser = {
-        "platform": "Windows 10",
-        "browserName": "firefox",
-        "version": "49.0"
-    }
-
-    desired_caps.update(browser)
-    test_name = request.node.name
-    build_tag = environ.get('BUILD_TAG', None)
-    tunnel_id = environ.get('TUNNEL_IDENTIFIER', None)
-    username = environ.get('SAUCE_USER', None)
-    access_key = environ.get('SAUCE_KEY', None)
-
-    selenium_endpoint = "http://{}:{}@ondemand.saucelabs.com:80/wd/hub".format(username, access_key)
-    desired_caps['build'] = build_tag
-    # we can move this to the config load or not, also messing with this on a test to test basis is possible :)
-    desired_caps['tunnelIdentifier'] = tunnel_id
-    desired_caps['name'] = test_name
-
-    executor = RemoteConnection(selenium_endpoint, resolve_ip=False)
-    browser = webdriver.Remote(
-        command_executor=executor,
-        desired_capabilities=desired_caps
-    )
-    yield browser
-
-    # This is specifically for SauceLabs plugin.
-    # In case test fails after selenium session creation having this here will help track it down.
-    # creates one file per test non ideal but xdist is awful
-    if browser:
-        with open("{}.testlog".format(browser.session_id), 'w') as f:
-            f.write("SauceOnDemandSessionID={} job-name={}\n".format(browser.session_id, test_name))
-    else:
-        raise WebDriverException("Never created!")
-
-    def fin():
-        browser.execute_script("sauce:job-result={}".format(str(not request.node.rep_call.failed).lower()))
-        browser.quit()
-    request.addfinalizer(fin)
 
 @pytest.mark.usefixtures('driver')
 class TestLink:
